Don't rely on an id field.
There is no guarantee that a model has an `id` field.

Instead, it's better to use the `.pk`, as this is guaranteed to be present.

Likewise, it is possible to have non-integer primary key columns. 

If you do have an integer primary key, you can pass a string (that contains a valid integer), and that will query fine.